Detect the default group
This commit changes this way the group parameter works.

Previously, the group parameter had a default of 'default'. However, in
many cases, the default resource group is not named 'default', or the
api-key may not have access to the resource group named 'default'.

After this commit, the default value for the group paramter is unset,
which will trigger a post-login detection and targeting of the actual
default resource group. For users who don't want this behavior, they can
pass `false` to the group parameter, and the detection will be skipped.
As before, if any other value is passed that value will be used as the
group.

Signed-off-by: Jason Frey <fryguy9@gmail.com>